requirements.txt
# Core dependencies
google-cloud-firestore>=2.13.0
google-auth>=2.23.0
huggingface_hub>=0.20.0

# Data processing
pandas>=2.0.0
pydantic>=2.0.0

# Readability metrics
textstat>=0.7.3

# Web framework
flask>=3.0.0

# Environment variables
python-dotenv>=1.0.0

